+++ kite-core/src/main/kotlin/com/tang/kite/config/optimistic/OptimisticLockConfig.kt
@@ -17,7 +17,7 @@ object OptimisticLockConfig {
      * The name of the version field
      */
     @JvmStatic
-    var versionFieldName: String = "version"
+    var fieldName: String = "version"
 
     /**
      * Default initial version value